nestedserializer/requirement.txt
amqp==2.6.0
asgiref==3.2.3
automium==0.2.6
automium-web==0.1.1
backtrader==1.9.75.123
billiard==3.6.3.0
celery==4.4.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
Django==2.2.2
django-crispy-forms==1.8.1
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
Faker==4.0.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
geoip2==3.0.0
graphviz==0.13.2
h5py==2.10.0
httpie==2.0.0
IbPy2==0.8.0
idna==2.8
importlib-metadata==1.6.1
inflection==0.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==4.6.10
legacy==0.1.6
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
maxminddb==1.5.2
more-itertools==8.3.0
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.4.6
numpy==1.18.5
pandas==1.0.4
passlib==1.7.2
Pillow==6.2.1
pipenv==2018.11.26
protobuf==3.10.0
psycopg2==2.8.5
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
Quandl==3.5.0
requests==2.22.0
scipy==1.4.1
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.8
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
Werkzeug==0.16.0
xlrd==1.2.0
zipp==3.1.0
